added the refresh endpoint
