+++ services/auth/schemas.py
@@ -9,7 +9,7 @@ from pydantic import (
 from services.auth.password_service import validate_password_strength
 
 
-class SignupResponse(BaseModel):
+class AuthResponse(BaseModel):
 	message: str
 
 
@@ -23,3 +23,17 @@ class SignupRequest(BaseModel):
 	@classmethod
 	def validate_password(cls, value: str) -> str:
 		return validate_password_strength(value)
+
+
+class LoginRequest(BaseModel):
+	email: EmailStr
+	password: str
+
+	@field_validator('password')
+	@classmethod
+	def validate_password(cls, value: str) -> str:
+		return validate_password_strength(value)
+
+
+class RefreshRequest(BaseModel):
+	refresh_token: str
